Remove 'npm install' from Dockfile.dev
The development environment relies being attached to the working
directory to access the project. This means 'npm install' during
construction of the Dev environment will never work as it won't be
able to access the right information.

Add 'npm install' to Makefile when starting the Dev environment as we
have attached the codebase by then.