Switch to yarn package manager
It seemed to that yarn was easier to fix and pin dependency versions
when debugging issues. yarn also provides integrity checking and
caching when repeating yarn commands.

Based on a good experience so far, lets see if yarn can help the project
be a bit more hands-off to get things up and running.